# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: antsxdp2.0 admin  接口
  version: 2.0.0
  title: antsxdp 2.0 -admin
basePath: /antsxdp
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/role/list:
    post:
      tags:
        - 角色管理
      summary: 角色列表
      produces:
        - application/json
      parameters:
        - name: body
          description: map
          in: body
          type: string
          required: true
          default: "{\"page\":\"1\",\"limit\":\"10\",\"roleName\":\"\"}"
      responses:
        '200':
          description: 返回角色列表
          schema:
            $ref: '#/definitions/SysRoleEntityList'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - 角色管理
      summary: 获取角色信息
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              role:
                $ref: '#/definitions/SysRoleEntity'
  /sys/role/save:
    post:
      tags:
        - 角色管理
      summary: 添加角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
        - 角色管理
      summary: 修改角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
        - 角色管理
      summary: 删除角色
      produces:
        - application/json
      parameters:
        - name: body
          description: ID列表
          in: body
          type: string
          default: "{\"ids\":\"1,2,3\"}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 菜单管理
      summary: 获取菜单信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              menu:
                description: 菜单信息
                $ref: '#/definitions/SysMenuEntity'
  /sys/menu/save:
    post:
      tags:
        - 菜单管理
      summary: 添加菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - 菜单管理
      summary: 修改菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete/{menuId}:
    post:
      tags:
        - 菜单管理
      summary: 删除菜单
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/cdnsys/ip/control/list:
    post:
      tags:
        - IP管理
      summary: ip列表
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: "control:-1=所有,0=未定义,1=白名单（7层）,2=7层黑名单,3=3层黑名单"
          default: "{\"parentId\":0,\"page\":1,\"limit\":10,\"control\":-1,\"ip\":\"\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/ip/control/save:
    post:
      tags:
        - IP管理
      summary: ip（分组） 保存与修改
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description:
          default: "{\"id\":0,\"parentId\":0,\"ip\":\"1.1.1.1\",\"control\":1,\"status\":0,\"remark\":\"\",,\"ips\":\"1.1.1.1,1.2.2.2\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/certify/list:
    post:
      tags:
        - 证书申请管理
      summary: 申请列表
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: "status 0==待签发 1=签发成功 2=签发失败 "
          default: "{\"page\":1,\"limit\":10,\"user\":\"\",\"key\":\"\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/certify/save:
    post:
      tags:
        - 证书申请管理
      summary: 保存
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: "创建与修改name 与 remark"
          default: "{\"id\":0,\"commonName\":\"www.cdn.com\",\"remark\":\"cdn\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/certify/batDelete:
    post:
      tags:
        - 证书申请管理
      summary: 删除
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: ""
          default: "{\"ids\":\"1,2,3\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/certify/apply/certificate:
    post:
      tags:
        - 证书申请管理
      summary: 申请
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: ""
          default: "{\"siteId\":0,\"mode\":0}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/certify/get/certify/by/host:
    get:
      tags:
        - 证书申请管理
      summary: 根据host获取证书
      produces:
        - application/json
      parameters:
        - name: host
          in: query
          type: string
          description: ""
          default: "1.com"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysRoleEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysRoleEntity'
  SysRoleEntityEdit:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
  SysMenuEntity:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      parentName:
        description: 父菜单名称
        type: string
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...